Relocate word files to view folder where they are used
