+++ src/components/category-view/category-view.tsx
@@ -3,8 +3,8 @@ import { Link, useParams } from 'react-router-dom';
 import styles from './category-view.module.css';
 import fetchCategoryData from '../../services/categoriesApiService';
 import { Fragment, useContext, useEffect, useState } from 'react';
-import WordCardDesktop from '../dictionary-page/word-card-desktop';
-import WordCardMobile from '../dictionary-page/word-card-mobile';
+import WordCardDesktop from '../dictionary-view/word-card-desktop';
+import WordCardMobile from '../dictionary-view/word-card-mobile';
 import { DictionaryType, FvCategory, FvWord } from '../common/data';
 import MultiSwitch from '../common/multi-switch/multi-switch';
 import fetchWordsData from '../../services/wordsApiService';
