+++ src/components/dictionary-view/word-categories.tsx
